Add files via upload
This program finds the number of palindromic substrings in a given string
The string itself may be palindrome or not
It also outputs the longest palindromic substring(s) in the given string.
NB: This program considers a single character as palindromic to itself. E.g. "a" is palindromic to itself.
Enjoy and feel free to comment or update or ask questions